separate service interface and implementation
